archive/directories/apps/backend/syncservice/pyproject.toml
[tool.poetry]
name = "syncservice"
version = "0.1.0"
description = "SyncService plugin for TerraFusion that migrates data from legacy PACS to CAMA systems"
authors = ["TerraFusion Team <info@terrafusion.io>"]
readme = "README.md"
packages = [{include = "syncservice", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.23"
pyodbc = "^4.0.39"
psycopg2-binary = "^2.9.9"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
nats-py = "^2.6.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
aiodataloader = "^0.3.0"
asyncio = "^3.4.3"
tenacity = "^8.2.3"
openai = "^1.1.1"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.1"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"
flake8 = "^6.1.0"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
